requirements.txt
Django==3.0.7
Pillow==7.1.2
asgiref==3.2.7
dj-database-url==0.5.0
django-on-heroku==1.0.1
django-rest-framework==0.1.0
django-storages==1.9.1
dropbox==10.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
